+++ src/forms/mod.rs
@@ -1,5 +1,7 @@
 pub(crate) mod agreement;
 pub(crate) mod cloud;
+pub mod cloud_firewall;
+pub mod firewall;
 pub mod project;
 pub mod rating;
 pub mod remote_secret;
@@ -8,6 +10,8 @@ pub mod status_panel;
 pub mod user;
 
 pub use cloud::*;
+pub use cloud_firewall::*;
+pub use firewall::*;
 pub use remote_secret::*;
 pub use server::*;
 pub use user::UserForm;