+++ src/startup.rs
@@ -400,6 +400,7 @@ pub async fn run(
                     .service(crate::routes::server::secret::item)
                     .service(crate::routes::server::secret::upsert)
                     .service(crate::routes::server::secret::delete)
+                    .service(crate::routes::server::cloud_firewall::configure)
                     .service(crate::routes::server::ssh_key::generate_key)
                     .service(crate::routes::server::ssh_key::upload_key)
                     .service(crate::routes::server::ssh_key::get_public_key)